dataAnalyst-agent/pyproject.toml
[tool.poetry]
name = "dataanalyst-agent"
version = "0.1.0"
description = ""
authors = ["Sagar <contact.sagargr@gmail.com>"]
readme = "README.md"
packages = [{include = "dataanalyst_agent"}]

[tool.poetry.dependencies]
python = "3.10.15"
urllib3 = "1.26.19"
llama-index = "^0.11.18"
llama-index-llms-openai = "^0.2.15"
llama-index-multi-modal-llms-openai = "^0.2.2"
llama-index-vector-stores-qdrant = "^0.3.2"
clip = {git = "https://github.com/openai/CLIP.git"}
langchain = "^0.3.4"
langchain-openai = "^0.2.3"
pillow = "10.2.0"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
llama-index-embeddings-clip = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
